refactor(forge): single context-aware studio at /forge/cards/[cardId] (+ legacy redirect)
