+++ app/forge/cards/[cardId]/FullModeForm.tsx
